HighScore and Final
Attempted to give the boss an ability to attack the player, however it was not able to do this. I added a high score function to the game that will only save the high score if the level is completed. I also completed the third level and have play tested it